backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
pydantic-settings>=2.3
python-dotenv>=1.0
requests>=2.31
